fix : 중복 dto 통일
